feature: 메시지 기본 셋팅
