+++ src/main/java/com/growith/tailo/chat/message/ChatConsumer.java
@@ -1,0 +1,12 @@
+package com.growith.tailo.chat.message;
+
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@RequiredArgsConstructor
+@Component
+public class ChatConsumer {
+
+
+}
