+++ src/main/java/com/growith/tailo/chat/message/ChatDTO.java
@@ -1,0 +1,5 @@
+package com.growith.tailo.chat.message;
+
+public record ChatDTO(MessageType type, String content, Long accountId, Long roomId) {
+
+}
