+++ src/main/java/com/growith/tailo/chat/message/ChatMessage.java
@@ -1,0 +1,33 @@
+package com.growith.tailo.chat.message;
+
+import com.growith.tailo.chat.room.ChatRoom;
+import com.growith.tailo.chat.member.ChatMember;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "chat_messages")
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class ChatMessage {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "chat_message_id")
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "sender_id")
+    private ChatMember sender;
+
+    @ManyToOne
+    @JoinColumn(name = "chat_room_id")
+    private ChatRoom chatRoom;
+
+    @Column(nullable = false)
+    private String content;
+}
