+++ src/main/java/com/growith/tailo/chat/message/MessageType.java
@@ -1,0 +1,7 @@
+package com.growith.tailo.chat.message;
+
+
+public enum MessageType {
+    CHAT,
+
+}
