--- src/main/java/com/growith/tailo/chat/message/entity/MessageType.java
@@ -1,5 +1,0 @@
-package com.growith.tailo.chat.message.entity;
-
-public enum MessageType {
-     TALK, // 추가 메시지 타입 작성 알림, 입장, 퇴장 ?
-}
