--- src/main/java/com/growith/tailo/chat/message/repository/ChatMessageRepository.java
@@ -1,9 +1,0 @@
-package com.growith.tailo.chat.message.repository;
-
-import com.growith.tailo.chat.message.entity.ChatMessage;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ChatMessageRepository extends JpaRepository<ChatMessage,Long> {
-}
