--- src/main/java/com/growith/tailo/chat/message/service/ChatMessageService.java
@@ -1,10 +1,0 @@
-package com.growith.tailo.chat.message.service;
-
-public class ChatMessageService {
-    // 디비 저장
-
-    // 조회
-
-    // 삭제
-
-}
