--- src/main/java/com/growith/tailo/chat/roommemeber/ChatMemberRepository.java
@@ -1,8 +1,0 @@
-package com.growith.tailo.chat.roommemeber;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ChatMemberRepository extends JpaRepository<ChatRoomMember, Long> {
-}
